FIR: move Java type conversion functions to a separate file
